+++ demo/klangdemo/dep.cljs
@@ -1,9 +1,0 @@
-(ns ^:figwheel-always
-  klangdemo.dep
-  (:require-macros
-   [klang.core :as macros :refer [trac!]]))
-
-(trac! :loading)
-
-(trac! :loaded)
-